Updated templates and readme to note conflicting routes
