--- examples/custom-server/src/app/api/test-get/route.ts
@@ -1,5 +1,0 @@
-import { NextResponse } from 'next/server'
-
-export async function GET(): Promise<NextResponse> {
-  return NextResponse.json({ success: true })
-}
