--- examples/custom-server/src/app/api/test-post/route.ts
@@ -1,5 +1,0 @@
-import { NextResponse } from 'next/server'
-
-export async function POST(): Promise<NextResponse> {
-  return NextResponse.json({ success: true })
-}
