+++ templates/ecommerce/src/app/next-api/preview/route.ts
@@ -3,6 +3,11 @@ import { redirect } from 'next/navigation'
 
 import { payloadToken } from '../../_api/token'
 
+/**
+ * The Next.js API routes can conflict with Payload's own routes if they share the same path
+ * To avoid this you can customise the path of Payload or the API route of Nextjs as we've done here
+ * See readme: https://github.com/payloadcms/payload/tree/main/templates/ecommerce#conflicting-routes
+ *  */
 export async function GET(
   req: Request & {
     cookies: {
