+++ templates/ecommerce/src/app/next-api/revalidate/route.ts
@@ -2,6 +2,11 @@ import { revalidateTag } from 'next/cache'
 import type { NextRequest } from 'next/server'
 import { NextResponse } from 'next/server'
 
+/**
+ * The Next.js API routes can conflict with Payload's own routes if they share the same path
+ * To avoid this you can customise the path of Payload or the API route of Nextjs as we've done here
+ * See readme: https://github.com/payloadcms/payload/tree/main/templates/ecommerce#conflicting-routes
+ *  */
 export async function GET(request: NextRequest): Promise<Response> {
   const collection = request.nextUrl.searchParams.get('collection')
   const slug = request.nextUrl.searchParams.get('slug')
