--- templates/website/src/app/api/exit-preview/route.ts
@@ -1,6 +1,0 @@
-import { draftMode } from 'next/headers'
-
-export async function GET(): Promise<Response> {
-  draftMode().disable()
-  return new Response('Draft mode is disabled')
-}
