+++ templates/website/src/payload/hooks/revalidatePage.ts
@@ -16,7 +16,7 @@ export const revalidatePage: AfterChangeHook = ({ doc, req }) => {
     try {
       url = formatAppURL({ doc })
       const res = await fetch(
-        `${process.env.PAYLOAD_PUBLIC_SERVER_URL}/api/revalidate?secret=${process.env.REVALIDATION_KEY}&revalidatePath=${url}`,
+        `${process.env.PAYLOAD_PUBLIC_SERVER_URL}/next-api/revalidate?secret=${process.env.REVALIDATION_KEY}&revalidatePath=${url}`,
       )
 
       if (res.ok) {
