+++ templates/website/src/payload/utilities/revalidate.ts
@@ -9,7 +9,7 @@ export const revalidate = async (args: {
 
   try {
     const res = await fetch(
-      `${process.env.PAYLOAD_PUBLIC_SERVER_URL}/api/revalidate?secret=${process.env.REVALIDATION_KEY}&collection=${collection}&slug=${slug}`,
+      `${process.env.PAYLOAD_PUBLIC_SERVER_URL}/next-api/revalidate?secret=${process.env.REVALIDATION_KEY}&collection=${collection}&slug=${slug}`,
     )
 
     if (res.ok) {
